refactor: handle flow navigation in NsFlow
